+++ apps/web/src/features/resume/preview/preview.shared.helpers.test.ts
@@ -6,7 +6,7 @@ import {
 	getPreviewCanvasScale,
 	getResumePreviewGapValue,
 	getScaledPreviewPageSize,
-} from "./preview.shared";
+} from "./preview.shared.utils";
 
 describe("getScaledPreviewPageSize", () => {
 	it("multiplies both dimensions by the scale", () => {
