+++ apps/web/src/features/resume/preview/preview.shared.tsx
@@ -2,6 +2,7 @@ import type { ResumeData } from "@reactive-resume/schema/resume/data";
 import type { CSSProperties } from "react";
 import { Spinner } from "@reactive-resume/ui/components/spinner";
 import { cn } from "@reactive-resume/utils/style";
+import { DEFAULT_PDF_PAGE_SIZE, getResumePreviewGapValue, getScaledPreviewPageSize } from "./preview.shared.utils";
 
 export type ResumePreviewProps = {
 	className?: string;
@@ -19,11 +20,6 @@ export type ResolvedResumePreviewProps = ResumePreviewProps & {
 	showPageNumbers: boolean;
 };
 
-export type PreviewPageSize = {
-	height: number;
-	width: number;
-};
-
 type ResumePreviewLoaderProps = Pick<ResumePreviewProps, "pageClassName" | "showPageNumbers"> & {
 	pageCount?: number;
 	pageGap?: CSSProperties["gap"];
@@ -31,34 +27,7 @@ type ResumePreviewLoaderProps = Pick<ResumePreviewProps, "pageClassName" | "show
 	pageScale?: number;
 };
 
-const PDF_PAGE_RENDER_SCALE = 4;
-const MAX_PREVIEW_CANVAS_PIXELS = 16_777_216; // 4096 * 4096
-export const DEFAULT_PDF_PAGE_SIZE: PreviewPageSize = {
-	height: 841.89,
-	width: 595.28,
-};
-
 // ponytail: normalizeResumePreviewProps deleted — defaults now live in ResumePreview destructuring
-
-export const getPreviewCanvasScale = (width: number, height: number) => {
-	const devicePixelRatio = typeof window === "undefined" ? 1 : window.devicePixelRatio || 1;
-	const desiredScale = Math.max(PDF_PAGE_RENDER_SCALE, devicePixelRatio);
-	const desiredPixels = width * height * desiredScale * desiredScale;
-
-	if (desiredPixels <= MAX_PREVIEW_CANVAS_PIXELS) return desiredScale;
-
-	return Math.sqrt(MAX_PREVIEW_CANVAS_PIXELS / (width * height));
-};
-
-export const getScaledPreviewPageSize = (pageSize: PreviewPageSize, pageScale: number): PreviewPageSize => ({
-	height: pageSize.height * pageScale,
-	width: pageSize.width * pageScale,
-});
-
-export const getResumePreviewGapValue = (pageGap: CSSProperties["gap"]) =>
-	typeof pageGap === "number" && pageGap !== 0 ? `${pageGap}px` : pageGap;
-
-export const getResumePreviewPageCount = (data?: ResumeData) => Math.max(1, data?.metadata.layout.pages.length ?? 1);
 
 export function ResumePreviewLoader({
 	pageCount = 1,
