+++ apps/web/src/routes/builder/$resumeId/-sidebar/left/sections/profiles.test.tsx
@@ -1,5 +1,6 @@
 // @vitest-environment happy-dom
 
+import type { Resume } from "@/features/resume/builder/draft";
 import { render, screen } from "@testing-library/react";
 import { beforeAll, describe, expect, it, vi } from "vitest";
 import { i18n } from "@lingui/core";
@@ -41,11 +42,12 @@ type SectionItemProps = {
 };
 
 vi.mock("@/features/resume/builder/draft", () => ({
-	useCurrentResume: () => ({
-		data: {
-			sections: { profiles: { title: "Profiles", columns: 1, hidden: false, items: sectionItems } },
-		},
-	}),
+	useCurrentBuilderResumeSelector: (selector: (resume: Resume) => unknown) =>
+		selector({
+			data: {
+				sections: { profiles: { title: "Profiles", columns: 1, hidden: false, items: sectionItems } },
+			},
+		} as unknown as Resume),
 	useUpdateResumeData: () => vi.fn(),
 }));
 vi.mock("../shared/section-base", () => ({
