+++ apps/web/src/routes/builder/$resumeId/-sidebar/right/sections/statistics.tsx
@@ -8,32 +8,11 @@ import { Alert, AlertDescription, AlertTitle } from "@reactive-resume/ui/compone
 import { cn } from "@reactive-resume/utils/style";
 import { orpc } from "@/libs/orpc/client";
 import { SectionBase } from "../shared/section-base";
+import { computeDelta, getSparklinePoints } from "./statistics.utils";
 
 // Fetch 60 days so we can render a 30-day sparkline and compare it against the prior 30 days.
 const TREND_DAYS = 60;
 const WINDOW = 30;
-
-// Percent change of the most recent `window` days vs the `window` days before it.
-// Returns null when the prior period had no activity (division by zero / no baseline).
-export function computeDelta(series: number[], window: number): number | null {
-	const recent = series.slice(-window);
-	const previous = series.slice(-window * 2, -window);
-	const recentSum = recent.reduce((sum, n) => sum + n, 0);
-	const previousSum = previous.reduce((sum, n) => sum + n, 0);
-	if (previousSum === 0) return null;
-	return Math.round(((recentSum - previousSum) / previousSum) * 100);
-}
-
-// Polyline points for the sparkline, or null for degenerate inputs (fewer than two
-// points, or an all-zero series) where there is nothing meaningful to draw.
-export function getSparklinePoints(values: number[], width: number, height: number): string | null {
-	if (values.length < 2 || values.every((n) => n === 0)) return null;
-	const max = Math.max(...values, 1);
-	const step = width / (values.length - 1);
-	return values
-		.map((value, index) => `${(index * step).toFixed(1)},${(height - (value / max) * height).toFixed(1)}`)
-		.join(" ");
-}
 
 export function StatisticsSectionBuilder() {
 	const params = useParams({ from: "/builder/$resumeId" });
